separate class and mapping run tables, establish runmain wrappers for run parameter retrieval. move raw classification wrappers from utilities pipeline to utilities views.
