+++ pathogen_identification/management/commands/submit_televir_sample_metagenomics_run.py
@@ -17,11 +17,11 @@ from pathogen_identification.models import (
 )
 from pathogen_identification.utilities.tree_deployment import TreeProgressGraph
 from pathogen_identification.utilities.utilities_pipeline import (
-    RawReferenceUtils,
     SoftwareTreeUtils,
     Utils_Manager,
 )
 from pathogen_identification.utilities.utilities_views import (
+    RawReferenceUtils,
     calculate_reports_overlaps,
     set_control_reports,
 )
