+++ static/css/televir_sample_main.css
@@ -37,6 +37,10 @@
     width: 100px;
 }
 
+.progress_display {
+    width: 100px;
+}
+
 .nmapped {
     width: 130px;
 }
