+++ templates/pathogen_identification/sample_detail.html
@@ -382,6 +382,7 @@
                 </div>
             </div>
         </div>
+        
         {% if sort_performed %}
         <div id="summary" class="collapsible">
             <div class="card-block collapsible-summary" style="display: none;">
