+++ tests_televir/pathogen_identification_coverage_report.txt
@@ -97,35 +97,35 @@ pathogen_identification/migrations/0049_auto_20240523_1208.py                   
 pathogen_identification/migrations/0050_auto_20240603_1533.py                                    5      0   100%
 pathogen_identification/migrations/0051_rawreferencecompoundmodel_screening_count.py             5      0   100%
 pathogen_identification/migrations/__init__.py                                                   0      0   100%
-pathogen_identification/models.py                                                              891    320    64%
+pathogen_identification/models.py                                                              891    310    65%
 pathogen_identification/modules/assembly_class.py                                              240    145    40%
 pathogen_identification/modules/classification_class.py                                        598    313    48%
 pathogen_identification/modules/metadata_handler.py                                            358    147    59%
 pathogen_identification/modules/object_classes.py                                              966    289    70%
 pathogen_identification/modules/preprocess_class.py                                            262    177    32%
 pathogen_identification/modules/remap_class.py                                                1034    300    71%
-pathogen_identification/modules/run_main.py                                                    768    262    66%
+pathogen_identification/modules/run_main.py                                                    767    262    66%
 pathogen_identification/tables.py                                                              701    410    42%
 pathogen_identification/templatetags/custom_tags.py                                              6      3    50%
 pathogen_identification/templatetags/html_tags.py                                               86     65    24%
 pathogen_identification/templatetags/report_colors.py                                          109     83    24%
-pathogen_identification/tests.py                                                               793     50    94%
+pathogen_identification/tests.py                                                               811     46    94%
 pathogen_identification/urls.py                                                                  5      0   100%
 pathogen_identification/utilities/clade_objects.py                                              50     13    74%
 pathogen_identification/utilities/entrez_wrapper.py                                            229    124    46%
 pathogen_identification/utilities/mapping_flags.py                                              60     36    40%
-pathogen_identification/utilities/overlap_manager.py                                           562    229    59%
-pathogen_identification/utilities/phylo_tree.py                                                 87     25    71%
+pathogen_identification/utilities/overlap_manager.py                                           562    224    60%
+pathogen_identification/utilities/phylo_tree.py                                                 87     19    78%
 pathogen_identification/utilities/reference_utils.py                                           377    326    14%
 pathogen_identification/utilities/televir_bioinf.py                                            137    100    27%
-pathogen_identification/utilities/televir_parameters.py                                        107     21    80%
-pathogen_identification/utilities/tree_deployment.py                                           788    471    40%
+pathogen_identification/utilities/televir_parameters.py                                        107     12    89%
+pathogen_identification/utilities/tree_deployment.py                                           786    470    40%
 pathogen_identification/utilities/update_DBs.py                                                498    242    51%
 pathogen_identification/utilities/update_DBs_tree.py                                           528    480     9%
-pathogen_identification/utilities/utilities_general.py                                         239    123    49%
+pathogen_identification/utilities/utilities_general.py                                         239    117    51%
 pathogen_identification/utilities/utilities_pipeline.py                                       1641    608    63%
 pathogen_identification/utilities/utilities_televir_dbs.py                                     136     53    61%
-pathogen_identification/utilities/utilities_views.py                                           762    575    25%
-pathogen_identification/views.py                                                              1256   1063    15%
+pathogen_identification/utilities/utilities_views.py                                           762    380    50%
+pathogen_identification/views.py                                                              1254   1061    15%
 ----------------------------------------------------------------------------------------------------------------
-TOTAL                                                                                        19568  11511    41%
+TOTAL                                                                                        19581  11273    42%
